allow float for --package parameter

+++ lib/time_sheet/time/cmd.rb
@@ -129,7 +129,7 @@ class TimeSheet::Time::Cmd
       o.boolean '-v', '--verbose', 'be more verbose'
       o.boolean '--debug', 'drop into a REPL on errors'
       o.separator "\n  invoice options:"
-      o.integer '--package', 'for invoice output: build packages of this duration in hours', default: 0
+      o.float '--package', 'for invoice output: build packages of this duration in hours', default: 0.0
       o.integer '--petty', 'fold records under a certain threshold into a "misc" activity', default: 0
     end
   end
